+++ app/leaderboard/clan-card.tsx
