+++ app/leaderboard/clan-detail.dialog.tsx
