+++ app/leaderboard/dividend-vault-widget.tsx
