+++ app/leaderboard/sparkline.tsx
