+++ app/leaderboard/user-rankings-chart.tsx
