+++ next.config.mjs
@@ -9,6 +9,15 @@ const nextConfig = {
   images: {
     unoptimized: true,
   },
-}
+  redirects: async () => {
+    return [
+      {
+        source: "/",
+        destination: "/leaderboard",
+        permanent: true,
+      },
+    ];
+  },
+};
 
-export default nextConfig
+export default nextConfig;
